feat: consolidating the cleanHeaders library into the new fetch lib
