+++ src/cmds/openapi.js
@@ -5,9 +5,9 @@ const { prompt } = require('enquirer');
 const OASNormalize = require('oas-normalize');
 const promptOpts = require('../lib/prompts');
 const APIError = require('../lib/apiError');
-const { cleanHeaders } = require('../lib/cleanHeaders');
 const { getProjectVersion } = require('../lib/versionSelect');
 const fetch = require('../lib/fetch');
+const { cleanHeaders } = require('../lib/fetch');
 const FormData = require('form-data');
 const parse = require('parse-link-header');
 const { file: tmpFile } = require('tmp-promise');
