+++ src/lib/fetch.js
@@ -28,3 +28,26 @@ module.exports.handleRes = async function handleRes(res) {
   }
   return body;
 };
+
+/**
+ * Returns the basic auth header and any other defined headers for use in node-fetch API calls.
+ *
+ * @param {string} key The ReadMe project API key
+ * @param {Object} inputHeaders Any additional headers to be cleaned
+ * @returns An object with cleaned request headers for usage in the node-fetch requests to the ReadMe API.
+ */
+module.exports.cleanHeaders = function cleanHeaders(key, inputHeaders = {}) {
+  const encodedKey = Buffer.from(`${key}:`).toString('base64');
+  const headers = {
+    Authorization: `Basic ${encodedKey}`,
+  };
+
+  Object.keys(inputHeaders).forEach(header => {
+    // For some reason, node-fetch will send in the string 'undefined'
+    // if you pass in an undefined value for a header,
+    // so that's why headers are added incrementally.
+    if (typeof inputHeaders[header] === 'string') headers[header] = inputHeaders[header];
+  });
+
+  return headers;
+};
